// src/types.d.ts
export interface LoggerConfiguration {
	defaultLogChannel?: Snowflake;
	errorLogChannel?: Snowflake;
	logFilePath?: string;
	channelMap?: {
		[key: string]: Snowflake;
	};
	channelFunction?: (logType: string) => string;
}

/**
 * The Logger interface, it handles logging including error logging
 */
export interface Logger {
	readonly configuration: LoggerConfiguration;
	readonly client: Client;

	/**
	 * Log the given content or embed to the channel of the given LogType
	 */
	log: (type: string, data: string | EmbedBuilder) => Promise<void>;
	/**
	 * Log the given error
	 */
	logError: (error: unknown) => Promise<void>;
	/**
	 * Log the given string, embed title or embed description to a file
	 */
	logToFile: (type: string, data: string | EmbedBuilder) => Promise<void>;
}